Medical_Waste_app/requirement.txt
streamlit
tensorflow
pillow
numpy
matplotlib
pandas
scikit-learn
seaborn
